+++ test/functional/retry.js
@@ -1,6 +1,7 @@
 'use strict';
 
 import cudl from './../../index';
+import should from 'should';
 import nock from 'nock';
 
 
@@ -10,7 +11,7 @@ describe('Retry request', () => {
 
         nock('http://localhost')
             .post('/')
-            .twice()
+            .times(2)
             .reply(500);
 
         nock('http://localhost')
@@ -20,15 +21,33 @@ describe('Retry request', () => {
         cudl.post
             .to('http://localhost')
             .max_retry(3)
-            .then((err, result, request) => {
+            .then((err) => {
 
-                if (err && err.code >= 500) {
-                    return request.retry();
-                }
+                should(err).be.eql(null);
 
                 done();
             });
+    });
 
+
+    it ('should reach max retry', done => {
+
+        nock('http://localhost')
+            .post('/')
+            .times(3)
+            .reply(500);
+
+        cudl.post
+            .to('http://localhost')
+            .max_retry(3)
+            .then((err) => {
+
+                err.max_retry_reached.should.be.exactly(true);
+                err.code.should.be.exactly(500);
+                err.errors.length.should.be.exactly(2);
+
+                done();
+            });
     });
 
 });