+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+v1.3.0
+----
+
+* Auto-retry for error 500s
+* Use the last error when max retry is reached
+
+
 v1.2.0
 ----